requirements.txt
astroid==2.11.4
click==8.1.3
dill==0.3.4
Flask==2.1.2
Flask-Login==0.6.1
Flask-WTF==1.0.1
importlib-metadata==4.11.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
pylint==2.13.8
pylint-flask==0.6
pylint-plugin-utils==0.7
python-dotenv==0.20.0
tomli==2.0.1
typing_extensions==4.2.0
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
